lambda/requirements.txt
# Lambda runtime dependencies
boto3>=1.34.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
pyarrow>=14.0.0
